requirements.txt
sortedcontainers
configobj
feedparser
eventlet
redis
pytz
python-socketio
Unidecode
flask
flask-socketio
requests
